Methods

Here are listed all the available methods of MegaQuiverTools

MegaQuiverTools.Quiver — Type

A quiver is represented by its adjacency $n \times n$ matrix $adjacency = (a_{ij})$, where $n$ is the number of vertices and $a_{ij}$ is the number of arrows $i \to j$.

Attributes:

adjacency is the adjacency matrix of the quiver

name is the name of the quiver, defaults to "".

MegaQuiverTools.Eulerform — Method

Computes the Euler form of the quiver for vectors x and y.

The Euler form is defined as the bilinear form

\[\langle x,y\rangle = x^T * E * y,\]

where E is the Euler matrix of the quiver.

MegaQuiverTools.Telemanbound_onstratum — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS $\lambda$ corresponding to the given HN type.

MegaQuiverTools.allHNtypes — Function

Returns a list of all the Harder Narasimhan types of representations of Q with dimension vector d, with respect to the slope function theta/slope_denominator.

MegaQuiverTools.allTelemanbounds — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS corresponding to each HN type for the given Q, d, theta and denominator.

MegaQuiverTools.all_generic_subdimension_vectors — Method

Returns the list of all generic subdimension vectors of d.

MegaQuiverTools.all_weights_endomorphisms_universal_bundle — Function

Computes the weights of the endomorphisms of the universal bundles $U_i \otimes U_j$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type.

MegaQuiverTools.all_weights_irreducible_component_canonical — Function

Computes the weights of the irreducible component of $\omega_R|_Z$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type. More explicitly, if $\omega_X = O(rH)$, this returns the weights of the pullback of O(H) on each stratum.

MegaQuiverTools.all_weights_universal_bundle — Function

Computes the weights of the universal bundle $U_i(a)$ for the linearization $a$ on all the non-dense Harder-Narasimhan strata for each 1-PS corresponding to each HN type.

MegaQuiverTools.canonical_stability — Method

The canonical stability parameter for the couple $(Q,d)$ is given by $<d,- > - < - ,d>$

MegaQuiverTools.codimensionHNstratum — Method

Returns the codimension of the given HN stratum.

MegaQuiverTools.euler_matrix — Method

Returns the Euler matrix of the quiver.

The Euler matrix of a quiver Q is defined as

\[E = I - A,\]

where $A$ is the adjacency matrix of Q and $I$ is the identity matrix of the same size as $A$.

MegaQuiverTools.hassemistables — Function

Checks if there is a theta-semistable representation of dimension vector d.

Examples:

julia> A2 = mKroneckerquiver(1)
julia> theta = [1,-1]
julia> d = [1,1]
julia> hassemistables(A2, d, theta)
true

julia> d = [2,2]
julia> hassemistables(A2, d, theta)
true

julia> d = [1,2]
julia> hassemistables(A2, d, theta)
false

julia> d = [0,0]
julia> hassemistables(A2, d, theta)
true

# The 3-Kronecker quiver:
julia> K3 = mKroneckerquiver(3)
julia> theta = [3,-2]
julia> d = [2,3]
julia> hassemistables(K3, d, theta)
true

julia> d = [1,4]
julia> hassemistables(K3, d, theta)
false

MegaQuiverTools.hasstables — Method

Checks if Q has a theta-stable representation of dimension vector d.

MegaQuiverTools.indegree — Method

Returns the number of incoming arrows to the vertex j.

Examples:

julia> Q = mKroneckerquiver(4)
julia> indegree(Q, 1)
0
julia> indegree(Q, 2)
4

MegaQuiverTools.isSchurroot — Method

Checks if d is a Schur root for Q. By a lemma of Schofield (See Lemma 4.2 of https://arxiv.org/pdf/0802.2147.pdf), this is equivalent to the existence of a stable representation of dimension vector d.

MegaQuiverTools.is_generic_subdimension_vector — Method

Checks if e is a generic subdimension vector of d.

A dimension vector e is called a generic subdimension vector of d if a generic representation of dimension vector d possesses a subrepresentation of dimension vector e.

By a result of Schofield (see Thm. 5.3 of https://arxiv.org/pdf/0802.2147.pdf) e is a generic subdimension vector of d if and only if $<e',d-e> \geq 0$ for all generic subdimension vectors e' of e.

MegaQuiverTools.isacyclic — Method

Checks wether the quiver is acyclic, i.e. has no oriented cycles.

MegaQuiverTools.isamplystable — Function

Checks wether the dimension vector d is amply stable with respect to the slope function theta/denominator.

This means that the codimension of the unstable locus in the parameter space is at least 2.

MegaQuiverTools.isconnected — Method

Checks wether the underlying graph of the quiver is connected.

Examples:

julia> Q = Quiver([0 1 0; 0 0 1; 1 0 0])
julia> isconnected(Q)
true

julia> Q = Quiver([0 1 0; 1 0 0; 0 0 2])
false

# The 4-Kronecker quiver:
julia> Q = mKroneckerquiver(4)
julia> isconnected(Q)
true

# The 4-loop quiver:
julia> Q = LoopQuiver(4)
julia> isconnected(Q)
true

# The 4-subspace quiver:
julia> Q = SubspaceQuiver(4)
julia> isconnected(Q)
true

# The A10 quiver:
julia> A10 = Quiver(   [0 1 0 0 0 0 0 0 0 0;
                        0 0 1 0 0 0 0 0 0 0;
                        0 0 0 1 0 0 0 0 0 0;
                        0 0 0 0 1 0 0 0 0 0;
                        0 0 0 0 0 1 0 0 0 0;
                        0 0 0 0 0 0 1 0 0 0;
                        0 0 0 0 0 0 0 1 0 0;
                        0 0 0 0 0 0 0 0 1 0;
                        0 0 0 0 0 0 0 0 0 1;
                        0 0 0 0 0 0 0 0 0 0] )
julia> isconnected(A10)
true

# The A10 quiver without one arrow:
julia> A10 = Quiver(   [0 1 0 0 0 0 0 0 0 0;
                        0 0 1 0 0 0 0 0 0 0;
                        0 0 0 1 0 0 0 0 0 0;
                        0 0 0 0 1 0 0 0 0 0;
                        0 0 0 0 0 1 0 0 0 0;
                        0 0 0 0 0 0 0 0 0 0;
                        0 0 0 0 0 0 0 1 0 0;
                        0 0 0 0 0 0 0 0 1 0;
                        0 0 0 0 0 0 0 0 0 1;
                        0 0 0 0 0 0 0 0 0 0] )
julia> isconnected(A10)
false

MegaQuiverTools.iscoprime — Method

Checks wether the given dimension vector $d$ is $theta$-coprime for the stability parameter $theta$.

MegaQuiverTools.iscoprime — Method

Checks if the gcd of all the entries of d is $1$.

MegaQuiverTools.issink — Method

Checks if the vertex j is a sink, i.e., a vertex with no outgoing arrows.

Examples:

julia> Q = mKroneckerquiver(4)
julia> issink(Q, 1)
false
julia> issink(Q, 2)
true

MegaQuiverTools.issource — Method

Checks if the vertex i is a source, i.e., a vertex with no incoming arrows.

Examples:

julia> Q = mKroneckerquiver(4)
julia> issource(Q, 1)
true
julia> issource(Q, 2)
false

MegaQuiverTools.narrows — Method

Returns the number of arrows of the quiver.

MegaQuiverTools.nvertices — Method

Returns the number of vertices of the quiver.

MegaQuiverTools.outdegree — Method

Returns the number of outgoing arrows from the vertex i.

Examples:

julia> Q = mKroneckerquiver(4)
julia> outdegree(Q, 1)
4
julia> outdegree(Q, 2)
0

MegaQuiverTools.underlying_graph — Method

Returns the (necessarily symmetric) adjacency matrix of the underlying graph of the quiver.

MegaQuiverTools.weight_irreducible_component_canonical_on_stratum — Function

Computes the weight of the irreducible component of $\omega_R|_Z$ on a Harder-Narasimhan stratum for the 1-PS corresponding to each HN type. More explicitly, if $\omega_X = O(rH)$, this returns the weight of the pullback of O(H) on the given stratum.

MegaQuiverTools.weights_endomorphism_universal_bundle_on_stratum — Function

Computes the weights of the endomorphism of the universal bundle $U_i \otimes U_j$ on the given Harder-Narasimhan stratum for the 1-PS relative to the HN type.

MegaQuiverTools.weights_universal_bundle_onstratum — Function

Returns the weights of a universal bundle $U_i(a)$ for the linearization $a$ for the 1-PS corresponding to the given HN type.